refactor(maestro): split ide/semantic_tokens into cohesive modules
